ml-fastapi/requirements.txt
fastapi
uvicorn
pandas
joblib
xgboost
googlemaps
python-dotenv
PyJWT